+++ src/main/java/com/eraky/crudSystemSample/Repository/OrdersRepository.java
@@ -3,5 +3,5 @@ package com.eraky.crudSystemSample.Repository;
 import com.eraky.crudSystemSample.entity.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrdersRepository extends JpaRepository<Order, Integer> {
+public interface OrdersRepository extends JpaRepository<Order, Long> {
 }
